added deleting and updating course, created page repository controller and service

+++ angular-JavaLearn/src/app/pages/admin/courseMaker/courseMaker.component.ts
@@ -1,6 +1,8 @@
 import { Component, OnInit } from '@angular/core';
+import { Router } from '@angular/router';
 import { Course } from 'src/app/models/course/Course';
 import { CourseService } from 'src/app/services/course.service';
+import { UtilsService } from 'src/app/services/utils.service';
 
 @Component({
   selector: 'app-courseMaker',
@@ -11,7 +13,7 @@ export class CourseMakerComponent implements OnInit {
 
   courseList: Array<Course> = []
   
-  constructor(private courseService: CourseService) {}  
+  constructor(private courseService: CourseService, private router: Router, private utils: UtilsService) {}  
   
   ngOnInit(): void {
     this.courseService.findAllCourses().subscribe( data => {
@@ -20,7 +22,10 @@ export class CourseMakerComponent implements OnInit {
   }
 
   delete(id: bigint) {
-    this.courseService.deleteCourse(id)
+    if(confirm("Are you sure to delete this course?")) {
+      this.courseService.deleteCourse(id)
+      this.utils.reloadComponent(this.router.url)
+    }
   }
 
 }
